Updated BOM & Schematic
BOM was too high.

Switched to a cheaper 8bit DAC. This will still yield ~0.06V resolution
when gained up to +15V.

Currently leaving power circuits alone, but in the future this could be
better optimized for space and cost. Multichannel buck PMIC would be a
good option to simplify the digital supplies.